cryptocurrency/requirements.txt
Django==4.2.11
djangorestframework
gunicorn
requests
